requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.11.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
cors==1.0.1
passlib[argon2]
python-jose[cryptography]
